moved tensor tests separately
