+++ hub/core/chunk_engine/tests/test_benchmark_chunk_engine.py
@@ -4,13 +4,14 @@ import numpy as np
 import pytest
 from hub.constants import GB
 
+from hub.util.array import get_random_array
 from hub.core.tensor import (
     read_samples_from_tensor,
     add_samples_to_tensor,
     create_tensor,
 )
 from hub.core.meta.tensor_meta import tensor_meta_from_array
-from hub.core.chunk_engine.tests.common import TENSOR_KEY, get_random_array
+from hub.core.tests.common import TENSOR_KEY
 from hub.core.typing import StorageProvider
 from hub.tests.common_benchmark import (
     parametrize_benchmark_chunk_sizes,
