+++ hub/core/tests/test_tensor_fixed_shape.py
@@ -2,20 +2,15 @@ from typing import Tuple
 
 import numpy as np
 import pytest
-from hub.constants import GB, MB
-from hub.core.chunk_engine.tests.common import get_random_array, run_engine_test
+from hub.core.tests.common import run_engine_test
+from hub.util.array import get_random_array
 from hub.tests.common import (
-    CHUNK_SIZE_PARAM,
-    CHUNK_SIZES,
-    DTYPE_PARAM,
-    DTYPES,
     SHAPE_PARAM,
     parametrize_chunk_sizes,
     parametrize_num_batches,
     parametrize_dtypes,
 )
 from hub.core.tests.common import (
-    parametrize_all_caches,
     parametrize_all_storages_and_caches,
 )
 from hub.core.typing import StorageProvider
