+++ hub/util/tests/test_array.py
@@ -2,8 +2,7 @@ import pytest
 
 import numpy as np
 
-from hub.util.array import normalize_and_batchify_shape
-from hub.core.chunk_engine.tests.common import get_random_array
+from hub.util.array import normalize_and_batchify_shape, get_random_array
 from hub.tests.common import parametrize_dtypes
 
 from typing import Tuple
